refactor(construction): split equipmentBVResolver into focused modules
Extract types, normalization patterns, catalog loading, and BV resolution logic
from equipmentBVResolver.ts (936 lines) into focused subdirectory modules:
- types.ts: BV result types, catalog entry types (53 lines)
- catalogLoader.ts: Catalog loading and caching (158 lines)
- normalizationPatterns.ts: Normalization patterns and alias maps (310 lines)
- normalization.ts: Equipment ID normalization logic (117 lines)
- resolution.ts: BV lookup and heat override logic (118 lines)
- catalogData.ts: Re-export from equipmentBVCatalogData.ts (13 lines)
- index.ts: Barrel export with public API (21 lines)

Original file reduced to facade re-export (1 line).
All modules under 400 lines. All public APIs preserved via barrel export.
No changes to BV resolution logic, heat override formulas, or catalog data values.
