+++ src/utils/construction/equipmentBV/catalogData.ts
@@ -1,0 +1,13 @@
+/**
+ * Equipment BV Resolver - Catalog Data Re-export
+ *
+ * Re-exports catalog data from equipmentBVCatalogData.ts for use by other modules.
+ */
+
+export {
+  AMMUNITION_CATALOG_FILES,
+  ELECTRONICS_CATALOG_FILES,
+  MISCELLANEOUS_CATALOG_FILES,
+  NAME_MAPPINGS_DATA,
+  WEAPON_CATALOG_FILES,
+} from '../equipmentBVCatalogData';
