+++ src/utils/construction/equipmentBV/normalizationPatterns.ts
@@ -1,0 +1,310 @@
+/**
+ * Equipment BV Resolver - Normalization Patterns
+ *
+ * Regex patterns and alias maps for normalizing equipment IDs from various formats
+ * (MegaMek, unit JSON, concatenated, etc.) to canonical catalog IDs.
+ */
+
+export const NORMALIZATION_PATTERNS: Array<[RegExp, string]> = [
+  [/^(?:is|cl)ac(\d+)$/, 'ac-$1'],
+  [/^(?:is|cl)lrm(\d+)$/, 'lrm-$1'],
+  [/^(?:is|cl)srm(\d+)$/, 'srm-$1'],
+  [/^(?:is|cl)?mrm(\d+)$/, 'mrm-$1'],
+  [/^(?:is|cl)?uac(\d+)$/, 'uac-$1'],
+  [/^(?:is|cl)?ultraac(\d+)$/, 'uac-$1'],
+  [/^(?:is|cl)?ultra-ac-(\d+)$/, 'uac-$1'],
+  [/^(?:is|cl)?lbxac(\d+)$/, 'lb-$1-x-ac'],
+  [/^(?:is|cl)?lb(\d+)xac$/, 'lb-$1-x-ac'],
+  [/^(?:is|cl)?lb-(\d+)-x-ac$/, 'lb-$1-x-ac'],
+  [/^(?:is|cl)?rac(\d+)$/, 'rac-$1'],
+  [/^(?:is|cl)?rotaryac(\d+)$/, 'rac-$1'],
+  [/^(?:is|cl)?lac(\d+)$/, 'lac-$1'],
+  [/^(?:is|cl)?lightac(\d+)$/, 'lac-$1'],
+  [/^(?:is|cl)?streaksrm(\d+)$/, 'streak-srm-$1'],
+  [/^(?:is|cl)?streak-srm-(\d+)$/, 'streak-srm-$1'],
+  [/^cl-?streaksrm(\d+)$/, 'clan-streak-srm-$1'],
+  [/^cl-?streak-srm-(\d+)$/, 'clan-streak-srm-$1'],
+  [/^(?:cl)?atm(\d+)$/, 'atm-$1'],
+  [/^(?:is)?mml(\d+)$/, 'mml-$1'],
+  [/^(?:is|cl)?smalllaser$/, 'small-laser'],
+  [/^(?:is|cl)?mediumlaser$/, 'medium-laser'],
+  [/^(?:is|cl)?largelaser$/, 'large-laser'],
+  [/^(?:is)?ersmall(?:laser)?$/, 'er-small-laser'],
+  [/^(?:is)?ermedium(?:laser)?$/, 'er-medium-laser'],
+  [/^(?:is)?erlarge(?:laser)?$/, 'er-large-laser'],
+  [/^(?:is)?er-small-laser$/, 'er-small-laser'],
+  [/^(?:is)?er-medium-laser$/, 'er-medium-laser'],
+  [/^(?:is)?er-large-laser$/, 'er-large-laser'],
+  [/^cl-?ersmall(?:laser)?$/, 'clan-er-small-laser'],
+  [/^cl-?ermedium(?:laser)?$/, 'clan-er-medium-laser'],
+  [/^cl-?erlarge(?:laser)?$/, 'clan-er-large-laser'],
+  [/^cl-?er-small-laser$/, 'clan-er-small-laser'],
+  [/^cl-?er-medium-laser$/, 'clan-er-medium-laser'],
+  [/^cl-?er-large-laser$/, 'clan-er-large-laser'],
+  [/^(?:is)?smallpulse(?:laser)?$/, 'small-pulse-laser'],
+  [/^(?:is)?mediumpulse(?:laser)?$/, 'medium-pulse-laser'],
+  [/^(?:is)?largepulse(?:laser)?$/, 'large-pulse-laser'],
+  [/^cl-?smallpulse(?:laser)?$/, 'clan-small-pulse-laser'],
+  [/^cl-?mediumpulse(?:laser)?$/, 'clan-medium-pulse-laser'],
+  [/^cl-?largepulse(?:laser)?$/, 'clan-large-pulse-laser'],
+  [/^(?:is)?ppc$/, 'ppc'],
+  [/^(?:is)?erppc$/, 'er-ppc'],
+  [/^(?:is)?er-ppc$/, 'er-ppc'],
+  [/^cl-?ppc$/, 'ppc'],
+  [/^cl-?erppc$/, 'clan-er-ppc'],
+  [/^cl-?er-ppc$/, 'clan-er-ppc'],
+  [/^(?:is|cl)?gaussrifle$/, 'gauss-rifle'],
+  [/^(?:is)?lightgaussrifle$/, 'light-gauss-rifle'],
+  [/^(?:is)?heavygaussrifle$/, 'heavy-gauss-rifle'],
+  [/^cl-?gaussrifle$/, 'clan-gauss-rifle'],
+  [/^(?:is|cl)?machinegun$/, 'machine-gun'],
+  [/^cl-?machinegun$/, 'clan-machine-gun'],
+  [/^(?:is)?flamer$/, 'flamer'],
+  [/^cl-?flamer$/, 'clan-flamer'],
+  [/^cl-?lrm(\d+)$/, 'clan-lrm-$1'],
+  [/^cl-?srm(\d+)$/, 'clan-srm-$1'],
+  [/^cl-?uac(\d+)$/, 'clan-uac-$1'],
+  [/^cl-?ultraac(\d+)$/, 'clan-uac-$1'],
+  [/^cl-?lbxac(\d+)$/, 'clan-lb-$1-x-ac'],
+  [/^cl-?lb(\d+)xac$/, 'clan-lb-$1-x-ac'],
+  [/^ac(\d+)$/, 'ac-$1'],
+  [/^lrm(\d+)$/, 'lrm-$1'],
+  [/^srm(\d+)$/, 'srm-$1'],
+];
+
+export const ABBREVIATION_MAP: Record<string, string> = {
+  slas: 'small-laser',
+  mlas: 'medium-laser',
+  llas: 'large-laser',
+  'er-slas': 'er-small-laser',
+  'er-mlas': 'er-medium-laser',
+  'er-llas': 'er-large-laser',
+  spl: 'small-pulse-laser',
+  mpl: 'medium-pulse-laser',
+  lpl: 'large-pulse-laser',
+};
+
+export const DIRECT_ALIAS_MAP: Record<string, string> = {
+  'ultra-ac-2': 'uac-2',
+  'ultra-ac-5': 'uac-5',
+  'ultra-ac-10': 'uac-10',
+  'ultra-ac-20': 'uac-20',
+  'clan-ultra-ac-2': 'clan-uac-2',
+  'clan-ultra-ac-5': 'clan-uac-5',
+  'clan-ultra-ac-10': 'clan-uac-10',
+  'clan-ultra-ac-20': 'clan-uac-20',
+  'rotary-ac-2': 'rac-2',
+  'rotary-ac-5': 'rac-5',
+  'rotary-ac-10': 'rac-10',
+  'clan-rotary-ac-2': 'clan-rac-2',
+  'clan-rotary-ac-5': 'clan-rac-5',
+  'hag-20': 'hag20',
+  'hag-30': 'hag30',
+  'hag-40': 'hag40',
+  'heavy-medium-laser': 'medium-heavy-laser',
+  'heavy-large-laser': 'large-heavy-laser',
+  'heavy-small-laser': 'small-heavy-laser',
+  'clan-heavy-medium-laser': 'medium-heavy-laser',
+  'clan-heavy-large-laser': 'large-heavy-laser',
+  'clan-heavy-small-laser': 'small-heavy-laser',
+  'autocannon-2': 'ac-2',
+  'autocannon-5': 'ac-5',
+  'autocannon-10': 'ac-10',
+  'autocannon-20': 'ac-20',
+  'light-ac-2': 'lac-2',
+  'light-ac-5': 'lac-5',
+  'light-auto-cannon-2': 'lac-2',
+  'light-auto-cannon-5': 'lac-5',
+  'rocket-launcher-1': 'rl1',
+  'rocket-launcher-2': 'rl2',
+  'rocket-launcher-3': 'rl3',
+  'rocket-launcher-4': 'rl4',
+  'rocket-launcher-5': 'rl-5',
+  'rocket-launcher-10': 'rl10',
+  'rocket-launcher-15': 'rl15',
+  'rocket-launcher-20': 'rl20',
+  'streak-lrm-1': 'streaklrm1',
+  'streak-lrm-2': 'streaklrm2',
+  'streak-lrm-3': 'streaklrm3',
+  'streak-lrm-4': 'streaklrm4',
+  'streak-lrm-5': 'streaklrm5',
+  'streak-lrm-6': 'streaklrm6',
+  'streak-lrm-7': 'streaklrm7',
+  'streak-lrm-8': 'streaklrm8',
+  'streak-lrm-9': 'streaklrm9',
+  'streak-lrm-10': 'streaklrm10',
+  'streak-lrm-11': 'streaklrm11',
+  'streak-lrm-12': 'streaklrm12',
+  'streak-lrm-13': 'streaklrm13',
+  'streak-lrm-14': 'streaklrm14',
+  'streak-lrm-15': 'streaklrm15',
+  'streak-lrm-16': 'streaklrm16',
+  'streak-lrm-17': 'streaklrm17',
+  'streak-lrm-18': 'streaklrm18',
+  'streak-lrm-19': 'streaklrm19',
+  'streak-lrm-20': 'streaklrm20',
+  'anti-missile-system': 'ams',
+  'clan-anti-missile-system': 'clan-ams',
+  'machine-gun-array': 'mga',
+  'light-machine-gun-array': 'lmga',
+  'heavy-machine-gun-array': 'hmga',
+  'clan-machine-gun-array': 'mga',
+  'clan-light-machine-gun-array': 'lmga',
+  'clan-heavy-machine-gun-array': 'hmga',
+  'arrow-iv': 'arrow-iv-launcher',
+  'clan-arrow-iv': 'clan-arrow-iv-launcher',
+  snppc: 'snub-nose-ppc',
+  'sn-ppc': 'snub-nose-ppc',
+  'er-micro-laser': 'clan-er-micro-laser',
+  'micro-pulse-laser': 'clan-micro-pulse-laser',
+  'er-flamer': 'flamer',
+  'clan-er-flamer': 'clan-flamer',
+  'c3-master-with-tag': 'c3-master',
+  'c3-computer-master': 'c3-master',
+  'c3-computer-slave': 'c3-slave',
+  'improved-c3-computer': 'c3i',
+  'small-x-pulse-laser': 'small-x-pulse-laser',
+  'medium-x-pulse-laser': 'medium-x-pulse-laser',
+  'large-x-pulse-laser': 'large-x-pulse-laser',
+  issmallxpulselaser: 'small-x-pulse-laser',
+  ismediumxpulselaser: 'medium-x-pulse-laser',
+  islargexpulselaser: 'large-x-pulse-laser',
+  'er-small-pulse-laser': 'clan-er-small-pulse-laser',
+  'er-medium-pulse-laser': 'clan-er-medium-pulse-laser',
+  'er-large-pulse-laser': 'clan-er-large-pulse-laser',
+  'small-chem-laser': 'clan-small-chemical-laser',
+  'medium-chem-laser': 'clan-medium-chemical-laser',
+  'large-chem-laser': 'clan-large-chemical-laser',
+  'particle-cannon': 'ppc',
+  'enhanced-ppc': 'enhanced-er-ppc',
+  'enhanced-er-ppc': 'enhanced-er-ppc',
+  'binary-laser-blazer-cannon': 'blazer-cannon',
+  'binary-laser-cannon': 'blazer-cannon',
+  'small-vsp-laser': 'small-vsp-laser',
+  'medium-vsp-laser': 'medium-vsp-laser',
+  'large-vsp-laser': 'large-vsp-laser',
+  'small-vsp': 'small-vsp-laser',
+  'medium-vsp': 'medium-vsp-laser',
+  'large-vsp': 'large-vsp-laser',
+  issmallvsplaser: 'small-vsp-laser',
+  ismediumvsplaser: 'medium-vsp-laser',
+  islargevsplaser: 'large-vsp-laser',
+  'medium-re-engineered-laser': 'medium-laser',
+  'large-re-engineered-laser': 'large-laser',
+  'small-re-engineered-laser': 'small-laser',
+  'srm-2-i-os': 'srm-2',
+  'srm-4-i-os': 'srm-4',
+  'srm-6-i-os': 'srm-6',
+  'lrm-5-i-os': 'lrm-5',
+  'lrm-10-i-os': 'lrm-10',
+  'lrm-15-i-os': 'lrm-15',
+  'lrm-20-i-os': 'lrm-20',
+  'heavy-ppc': 'heavy-ppc',
+  'light-ppc': 'light-ppc',
+  'hyper-velocity-ac-2': 'hyper-velocity-auto-cannon-2',
+  'hyper-velocity-ac-5': 'hyper-velocity-auto-cannon-5',
+  'hyper-velocity-ac-10': 'hyper-velocity-auto-cannon-10',
+  'hvac-2': 'hyper-velocity-auto-cannon-2',
+  'hvac-5': 'hyper-velocity-auto-cannon-5',
+  'hvac-10': 'hyper-velocity-auto-cannon-10',
+  isguardianecm: 'guardian-ecm',
+  isimprovedc3cpu: 'c3i',
+  isantimissilesystem: 'ams',
+  issnppc: 'snub-nose-ppc',
+  clheavymediumlaser: 'medium-heavy-laser',
+  'ismachine-gun': 'machine-gun',
+  isrotaryac5: 'rac-5',
+  isrotaryac2: 'rac-2',
+  clheavylargelaser: 'large-heavy-laser',
+  'protomech-ac-2': 'protomechac2',
+  'protomech-ac-4': 'protomechac4',
+  'protomech-ac-8': 'protomechac8',
+  'improved-heavy-gauss-rifle': 'improvedheavygaussrifle',
+  'tag-clan': 'clan-tag',
+  'light-tag-[clan]': 'clan-light-tag',
+  'lrt-5': 'lrm-5',
+  'lrt-10': 'lrm-10',
+  'lrt-15': 'lrm-15',
+  'lrt-20': 'lrm-20',
+  'srt-2': 'srm-2',
+  'srt-4': 'srm-4',
+  'srt-6': 'srm-6',
+  'clan-lrt-5': 'clan-lrm-5',
+  'clan-lrt-10': 'clan-lrm-10',
+  'clan-lrt-15': 'clan-lrm-15',
+  'clan-lrt-20': 'clan-lrm-20',
+  'clan-srt-2': 'clan-srm-2',
+  'clan-srt-4': 'clan-srm-4',
+  'clan-srt-6': 'clan-srm-6',
+  'anti-battlearmor-pods-b-pods': 'b-pod',
+  isbpod: 'b-pod',
+  clbpod: 'b-pod',
+  'c3-master-boosted-with-tag': 'c3-boosted-master',
+  islaserantimissilesystem: 'laser-ams',
+  cllaserantimissilesystem: 'clan-laser-ams',
+  isplasmarifle: 'plasma-rifle',
+  islppc: 'light-ppc',
+  isblazer: 'blazer-cannon',
+  iseherppc: 'enhanced-er-ppc',
+  clplasmacannon: 'clan-plasma-cannon',
+  clheavysmalllaser: 'small-heavy-laser',
+  clmicropulselaser: 'clan-micro-pulse-laser',
+  ismg: 'machine-gun',
+  clmg: 'clan-machine-gun',
+  islightmg: 'light-machine-gun',
+  issrm4os: 'srm-4',
+  isangelecm: 'angel-ecm',
+  clangelecmsuite: 'angel-ecm',
+  angelecmsuite: 'angel-ecm',
+  novacews: 'watchdog-cews',
+  watchdogecmsuite: 'watchdog-cews',
+  clantimissilesystem: 'clan-ams',
+  clactiveprobe: 'clan-active-probe',
+  cllightactiveprobe: 'light-active-probe',
+  clerflamer: 'clan-flamer',
+  iserflamer: 'flamer',
+  isc3mastercomputer: 'c3-master',
+  isc3slaveunit: 'c3-slave',
+  issniperartcannon: 'sniper-cannon',
+  sniper: 'sniper-cannon',
+  isarrowivsystem: 'arrow-iv-launcher',
+  clarrowivsystem: 'clan-arrow-iv-launcher',
+  isfluidgun: 'fluid-gun',
+  ismediumpulselaserprototype: 'medium-pulse-laser',
+  islbxac10prototype: 'lb-10-x-ac',
+  clrocketlauncher10prototype: 'rocket-launcher-10',
+  clrocketlauncher15prototype: 'rocket-launcher-15',
+  rocketlauncher20prototype: 'rocket-launcher-20',
+  'streak-srm-2-os': 'streak-srm-2',
+  'streak-srm-4-os': 'streak-srm-4',
+  'streak-srm-2-i-os': 'streak-srm-2',
+  'streak-srm-4-i-os': 'streak-srm-4',
+  'srm-2-os': 'srm-2',
+  'srm-6-os': 'srm-6',
+  'narc-i-os': 'narc',
+  'primitive-prototype-ppc': 'ppc',
+  ppcp: 'ppc',
+  'rifle-cannon': 'rifle-cannon-heavy',
+  'tsemp-one-shot': 'tsemp-cannon',
+  iselectronicwarfareequipment: 'electronic-warfare-ew-equipment',
+  'prototype-er-medium-laser': 'er-medium-laser',
+  'prototype-er-small-laser': 'er-small-laser',
+  'er-large-laser-prototype': 'er-large-laser',
+  'prototype-streak-srm-4': 'streak-srm-4',
+  'prototype-streak-srm-6': 'streak-srm-6',
+  'prototype-ultra-autocannon-10': 'uac-10',
+  'prototype-lb-10-x-autocannon': 'lb-10-x-ac',
+  'prototype-rocket-launcher-20': 'rocket-launcher-20',
+  'rocket-launcher-10-pp': 'rocket-launcher-10',
+  'ac-10p': 'ac-10',
+  'c3-boosted-system-master': 'c3-boosted-master',
+  'c3-computer-[master]': 'c3-master',
+  'c3-remote-sensor-launcher': 'c3-master',
+  isc3remotesensorlauncher: 'c3-master',
+  'plasma-cannon': 'clan-plasma-cannon',
+  'silver-bullet-gauss-rifle': 'sbgr',
+  taser: 'battlemech-taser',
+  'mech-taser': 'battlemech-taser',
+  ismektaser: 'battlemech-taser',
+  'flamer-vehicle': 'flamer',
+};
